Modify nvblox build for easier 0.0.8 integration.
- Bump project version from 0.0.7 -> 0.0.8 to match public release.
- Add `include(CTest)` and `enable_testing()` if BUILD_TESTING
is enabled to allow users to run `make test` directly (currently
users have to cd into the nvblox subfolder in the build dir and
run ctest).
- If user is using system eigen / stdgpu, don't install targets.
- Fix missing namespace declaration in `block_memory_pool.h`
- Fix missing end brace in `block_memory_pool_params.h`
- Fix ambiguity in `image_masker.cu` from fmax to std::fmax.
- Set working directory to `${CMAKE_CURRENT_SOURCE_DIR}` to allow
for datasets to be available to test executables when running `make test`.
- If using system eigen, set up `nvblox_eigen` target so the rest of
the build process proceeds as expected.
- If using system stdgpu, set up `nvblox_stdgpu` alias so the
rest of the build process works as expected.